feat(avro): use only native dart implementation
