--- lib/src/utils/implementations/bundle_tag_parser_ffi.dart
@@ -1,5 +1,0 @@
-import 'dart:typed_data';
-
-import 'package:arweave/arweave.dart';
-
-Uint8List serializeTags({required List<Tag> tags}) => throw UnimplementedError();
